# docs/openapi-rest.yaml
openapi: 3.0.3
info:
  title: LCP Platform API
  version: 1.1.0
  description: >
    REST and GraphQL surface for the self-hosted LCP platform. This document
    covers the admin dashboard login flow, publication processing, license
    lookup, status APIs, content downloads, publication catalog operations,
    runtime endpoints, and the GraphQL operations used by integration clients
    such as WordPress.
servers:
  - url: https://yourdomain.com
  - url: https://staging.yourdomain.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/licenses:
    get:
      summary: List licenses for the admin dashboard
      operationId: listAdminLicenses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: License list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLicensesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/licenses/{licenseId}:
    patch:
      summary: Update a license loan expiry date
      operationId: patchAdminLicense
      security:
        - bearerAuth: []
      parameters:
        - name: licenseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLicensePatchRequest"
      responses:
        "200":
          description: Updated license
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLicense"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Role or 2FA validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      additionalProperties: false
...
    AdminLicense:
      type: object
      properties:
        id:
          type: string
        publicationID:
          type: string
        userID:
          type: string
        passphrase:
          type: string
        hint:
          type: string
        publicationURL:
          type: string
        rightPrint:
          type: integer
        rightCopy:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminLicensesResponse:
      type: object
      properties:
        licenses:
          type: array
          items:
            $ref: "#/components/schemas/AdminLicense"
    AdminLicensePatchRequest:
      type: object
      properties:
        endDate:
          type: string
          format: date-time
...